Integrate Fortress as bnd bundle
The main idea is copied from how Khadija did it. There's a bnd bundle
org.alloytools.fortress.core that wraps a fortress submodule. The
build.gradle script calls sbt to build fortress, then copies the
resulting jar to a 'libs' directory and uses that as the bundle source.

Some hackery was required to get this to work in IntelliJ too, because
IntelliJ didn't seem to like the fact that the fortress bundle was just
a wrapper over a jar. I had to manually add the
ca.uwaterloo.watform.fortress.jar file as a dependency of the
org.alloytools.alloy.core module in Project Structure, then ignore the
warning that the bndtool extension gives about how fortress classes are
then imported from a non-bundle environment.