Create basic Portus architecture
It's based around the Translator interface, which is something that
tries to translate an Alloy Expr into a Fortress Theory. The
TranslatorManager then combines these Translators and for each Expr,
iterates through the translators until one translates successfully.

The idea is that we can implement each optimization as a translator
which intercepts certain Exprs before the default translator reaches it.
There's a convenience base implementation BaseTranslator to facilitate
this.