Handle oneOf in quantifier Decl lists
Previously we ignored ONEOF ExprUnary wrappers over quantifier terms in
the tests, but we can't do that with a whole file. It's also not
practical to translate "one A" on its own when it appears as part of a
decl, as "all x: one A". So we handle it in translateDeclList. We also
don't attempt to handle any of the other quantifiers (lone, some, set) -
perhaps in the future we'll try to skolemize like Kodkod does.