Bugfix: set univ's scope properly
We weren't using addToUnivScope at all. Oops. To make extra sure we
use a positive scope for the univ sort, we do a Math.min when setting it
in configureModelFinder.

Also, whenever we catch anything from Fortress, we rethrow it as an
ErrorFatal. Applications aren't *supposed* to catch Errors, but Alloy
does, and rethrowing Errors from Fortress like this gives a better debug
message.