Add translation for transitive and reflexive closure
We translate "(x,y) \in ^e" as "Closure(f(x,y))", where f:
(univ,univ)->Bool is an auxiliary defined via the axiom:
    forall x,y: univ . f(x,y) <=> [[(x,y) \in e]]
We have two mini-optimizations: if e already has a Fortress relation of
that form (i.e. it's a Sig.Field), we use it and skip the auxiliary
function, and if an auxiliary function has already been generated for e
according to isSame(), we use the old auxiliary function instead of
defining a new one.

Also, `./gradlew clean` should also clean Fortress now.