Large refactor: use SortPolicy, pass down sort info
SortPolicy is an abstraction over our policy for assigning sigs to
sorts. Currently we just assign everything to univ except Int
(UnivSortPolicy), but this can be expanded in the future.

To facilitate translating transitive closure, we also pass down sort
information for each variable in an ExprElementOf, using the VarTuple
abstraction.

We also do some short-circuiting: for many operations, if we have
incompatible sorts, we can declare the result to be false (Bottom)
rather than erroring. For example, x \in v where x and v have mismatched
types can be short-circuited to Bottom, because x can't be in v.