Implement sequences, seq/Int
It was surprisingly easy. Alloy implements "seq x" by interpreting it as
a function "seq/Int -> A", where seq/Int is a Kodkod builtin subsig of
Int containing only the ints in [0, maxseq-1]. (Also that arrow is
"isSeq->lone".)

We implement it by just translating "x \in seq/Int" as "0 <= x <=
maxseq-1" and enabling the function optimization. This results in
translating "seq A" as a function Int->A with a domain restriction.
Optimizations are probably possible on the Fortress level.