Support unchanging scopes
A sort is marked as 'unchanging' if its scope can't be changed in the
output without altering the semantics of the problem; that is, if you
change the scope in the SMTLIB+ output, the result doesn't represent the
same file. In practice this means that a sort is unchanging if we expand
over its domain elements in the Portus translation, meaning it's
involved in a sum or cardinality expression.