Handle simple integer join expressions
Sometimes join is used as an integer expression, like "x.size". This is
tricky because we don't want to try to support dealing with sets of
integers as integers (like Kodkod does - it just sums them), so we have
to handle this specially. Currently we only support "x.y" where x is a
bound variable or one sig (so it's a singleton set) and y is an
optimized function from sort(x) to Int.