Fix obscure bug in let mapping for calls
To translate a call like f[a, b, c] where is declared like
  pred f[x, y, z] { e }
we previously translated as
  let x=a | let y=b | let z=c | e
This works when all the variable names are distinct, but if the model
likes to reuse variable names (cough Chord cough) then the sequential
lets could interfere with each other - imagine that 'c' above referenced
a variable called 'x', then the above would use the wrong let mapping
for 'x'. The solution is to map all the arguments *simultaneously*,
which essentially just means using the same old let mapping for all of
them. We add some special support in TranslationContext to do this.