Add cardinality and quantifier scope axiom strategies
We use the strategy pattern for deciding the form we translate the scope
axiom in. The quantifier strategy (expressing a scope constraint with
quantifiers) tends to blow up on quantifier expansion, so we introduce
the cardinality strategy, which expresses scope constraints with
cardinality instead (like "#A <= 5"); this should grow like O(nlogn)
(where n is the scope) instead.