Fix overly restrictive 'in' sort restrictions
We were previously requiring both sides of an 'in' formula to have
definite sorts. This resulted in rejecting the commonly used
construction "f in iden" and variants. We now do it more intelligently
and allow indefinite sorts on the right-hand side of such 'in'
expressions. Indefinite sorts on the left-hand side and in '='
expressions are less intuitive and less common, so they're still
rejected, though if a need arises, short-circuiting could still be
implemented for those. See the comment in DefaultTranslator:572.