Handle domain element ranges for exact-scope sigs with non-exact parents
Turns out it was actually deeper than I thought. We don't support
ordering non-exact scopes, but we should support ordering exact-scope
children of non-exact scope sigs. So how do we determine the range of
domain elements to use? An exact-scope child of a non-exact-scope parent
forces the parent's scope to be a minimum size to accommodate the
child's scope, so we allocate that range of domain elements.