Bugfix: generate new closure aux function for different signatures
Previously we'd generate a new closure auxiliary function only if we
were closing over a different expression or the same expression with
different sorts. This works most of the time but there are edge cases:
consider
  forall v: S | ... ^(v->v) ...
vs
  let v = S | ... ^(v->v) ...
It's the same expression with the same sorts being closed over both
times, but there are different numbers of free variables, which was
generating an error like "Attempting to close over a relation that does
not end in a BoolSort or with the wrong argument sorts". Fixed by
generating a new aux relation when any part of its signature changes
(including the number of arguments), not just the result sort.