Bugfix: Sig.STRING getting -1 scope in getDomainElementRange
It turns out that ScopeComputer.sig2scope has the very annoying
behaviour of returning -1 whenever we don't know the scope of a sig. In
particular, when the scope of Sig.STRING isn't specified, we get -1.
This was messing up SortPolicy.getDomainElementRange since we didn't
handle it. To fix, since we don't support Sig.STRING yet, we just make
sure there are no strings and skip it.