Remove scope axioms from subset sigs
Turns out subset sigs can't have scopes specified. When we tried to
specify scopes for them, sig2scope would return -1, so we'd end up with
axioms of the form "#A <= -1", making the whole model UNSAT. Now we
don't generate scope axioms for subset sigs, and also we throw an
exception instead of silently failing if sig2scope returns -1 when
generating scope axioms.