Ignore non-exact sig atoms in getDomainElementRange
We had a bug with hierarchies like this:
     A,<=2
     /   \
A1,<=2   A2,=2
If we don't sort the exact-scope elements first, the domain element
range for A2 would be [3, 4], but there are only two domain elements in
the whole model!

To fix this, we essentially ignore non-exact-only domain elements from
domain element ranges. The first section of the domain element list is
then partitioned among the various exact-scope sigs. This has the result
that every exact-scope sig (which are the only sigs that can be ordered)
has a contiguous domain element range. Parent non-exact sorts have their
exact-scope children in a contiguous range.