Refactor getDomainElementRange to its own class, RangeAssigner
The algorithm really had nothing to do with SortPolicy and was
outgrowing that class. It'll need more complexity in the future, so
modularizing it out is a better choice. I don't really foresee any need
for variations on the algorithm, but if necessary RangeAssigner can be
made an interface to use the strategy pattern.