Only try values from sorts if known in evaluation
Previously we'd try all the values from all the sorts when evaluating an
expression. Now, if the sort policy can determine sorts for positions in
the expression, we only try values from those sorts (since the rest
should short-circuit to false). This speeds up evaluation dramatically
(on one example, we went from spending 88s to 2s on evaluation).

Also, cartesianProduct is lazy to avoid constructing huge lists. We ran
out of memory with an eager cartesianProduct on one example!