Refactor: support full Terms in VarTuple (now TermTuple)
A few times in the codebase, we had to do an ugly hack when we wanted to
use a complex term in an ExprElementOf where we'd use a dummy variable
and then substitute it. This became untenable, and also an upcoming
enhancement to the join optimization needs more than this. So, VarTuple
(now renamed TermTuple) contains AnnotatedTerms instead of
AnnotatedVars.

AnnotatedTerm contains a Term and a Sort (like AnnotatedVar), and also a
list of the AnnotatedVars which are free in the Term to support
retrieving a list of the free variables (see
PortusUtil.computeFreeVariables). This requires specifying the free
variables when creating an AnnotatedTerm, which as it turns out is easy.

Also, ContextVisitReturn is a FortressVisitReturn subclass which
abstracts handling keeping a TranslationContext updated when processing
nodes like quantifiers and automatically expanding let mappings. This
probably fixed some undiscovered bugs where not all this functionality
was implemented each time this pattern was used.