Refactor: generalize 'cast to scalar' out of FunctionOptTranslator
Previously castToScalar() was a function exclusively in
FunctionOptTranslator. This caused problems because it lead to some
unrelated optimizations being put there to use that function. Now we
generalize it to a similar chain-of-responsibility structure like
Translator. ScalarCaster is an interface defining a single
castToScalar() function, and implementations either return a (scalar,
guard) pair or null. TranslatorManager implements ScalarCaster and does
the same thing as for translators: it keeps a list of scalar casters and
returns the pair from the first one to return non-null.

This let us bring SimpleScalarOptTranslator out of
FunctionOptTranslator.