Extract VarMappingContext from TranslationContext
The functionality in TranslationContext that kept track of the mapping
from Alloy vars to Fortress Terms/let exprs was orthogonal to the rest
of the functionality in TranslationContext, but needed in PartitionSortPolicy,
and a SortPolicy is required to construct TranslationContext. So this
extracts that functionality to a new class, VarMappingContext.