Add Kodkod integer compatibility translator
Kodkod has these div-by-zero semantics: 0/0 = 0, x/0 = -1 for x>0,
x/0 = 1 for x<0. This is not guaranteed by Fortress. To allow full
compatibility for correctness checking, this adds
KodkodIntCompatibilityTranslator, which translates / in a manner that
maintains full compatibility. The translation is quite inefficient, so
it should not be used by default - it should only be used for
correctness checking against Kodkod. The command-line flag is
-kodkod-int-compat.