Bugfix: don't force exact scope when only children are subset sigs
It turns out that we don't need membership predicates when the only
children of a sig are subset sigs, but we were both forcing the scope to
be exact and not adding the membership predicate. This caused issues.
Now we don't force the scope to be exact in this case.

This bug was found by fuzz testing!