Bugfix: disable one sig opt for children of ordered sigs
The ordering module translator fixes a particular ordering of the domain
elements. The one sig optimization fixes particular domain elements for
the one sig. This can cause conflicts: e.g.
  open util/ordering[A]
  abstract sig A {}
  one sig A1, A2 extends A {}
  run { gt[A1, A2] }
This could fail due to this interaction. So we just disable the one sig
optimization for children of ordered signatures.