Don't add non-exact scope axiom when unnecessary
When the signature's size was the same as the sort's size but the sig
was non-exact, so the membership predicate optimization didn't trigger,
DefaultTranslator was still adding the non-exact scope axiom. E.g.:
  sig A {}
  run {} for 3 A
The non-exact scope axiom "#A <= 3" for the cardinality-based scope
axioms would still be added. This wasn't an issue with the
constants-based scope axioms but was with the cardinality-based scope
axioms. Now we don't add the non-exact scope axiom in DefaultTranslator
when using the membership predicate when the size of the sig equals the
size of the sort.