Fix bug with -scope
The error occurs when using -scope on sigs specified twice in the
command:
  sig A {}
  run {} for 3 A, 3 A
This seems like an oversight that it doesn't error out at all, but it
shows up in etl_scd.als in the expert models. -scope and -all-scopes was
changing just the first rather than both, causing an error. Now we
change both.