Mark scopes unchanging when necessary in ConstantsScopeAxiomStrategy
In ConstantsScopeAxiomStrategy, to bound the scope of a sig A within a
sort S to <=k, we generate |S|-k constants and assert they're all not in
A. This means S has to be unchanging (fixed) because it's no longer
semantically valid to change S's scope. This wasn't done before, now it
is.