Add extra arrow and join scalar casts
This adds castToScalar definitions for ~(e1->e2) where e1 is a scalar
with arity 1, and castToScalar(s.t) := castToScalar(t.~s) where s has
arity 2 and t has arity 1. The former recognizes that e1->e2 is a
bijection and the latter is for the off-chance that ~s is a unary scalar
function (like ~(e1->e2)) and t is a nilary scalar, which wouldn't be
able to use the regular composition castToScalar.