Bugfix: apply membership predicate opt in all possible cases
We previously were not applying the membership predicate optimization to
any sorts that are ever expanded over. However, this is only ever
required for non-exact top-level sigs when Fortress non-exact sorts are
in use! This is not being used for the paper, so we just don't apply
this condition when Fortress non-exact sorts are disabled.

The implementation still isn't optimal in the case that Fortress
non-exact sorts are used: we should refactor the ScopeExpansionMarkers
to return a list of sigs that are expanded over and only mark sorts
inapplicable if the corresponding sig is non-exact.