First implementation of 2nd order quantifiers
A variable that could be first or second-order is represented by a
DeclResult, which is basically a tagged union between AnnotatedVar and
FuncDecl. We *don't* support adding 2nd order variables to the
fortress-var stack with addFortressVar, so they won't show up as free
variables which will cause errors in some situations - this is not yet
implemented because getting free variables is primarily used for
abstracting formulas into definitions, and we don't (yet?) support 2nd
order arguments to definitions in Fortress.